Force 64-bit int in ProbeWDL

+++ fathom/constants.go
@@ -8,4 +8,4 @@ const (
 	TB_WIN
 )
 
-const TB_RESULT_FAILED = 0xFFFFFFFF
+const TB_RESULT_FAILED = int64(0xFFFFFFFF)
